// client/fastapi/models/RegisterUserDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisterUserDto = {
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    address: string;
    phone: string;
    dob: string;
};